backend/requirements.txt
Flask
Flask-Cors
supabase
python-dotenv
pyjwt
